chore: split platform utils
