+++ src/MeoAssistant/MeoAssistant.vcxproj.filters
@@ -402,6 +402,9 @@
     <ClInclude Include="SafeWindows.h">
       <Filter>头文件\Utils</Filter>
     </ClInclude>
+    <ClInclude Include="AsstPlatformWin32.h">
+      <Filter>头文件\Utils</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Controller.cpp">
@@ -674,7 +677,10 @@
     <ClCompile Include="ReportDataTask.cpp">
       <Filter>源文件\Task\Sub</Filter>
     </ClCompile>
-    <ClCompile Include="AsstUtils.cpp">
+    <ClCompile Include="AsstPlatformPosix.cpp">
+      <Filter>源文件</Filter>
+    </ClCompile>
+    <ClCompile Include="AsstPlatformWin32.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
   </ItemGroup>
